Striation Processor Nears Function
A near-function commit of the work being done on the Striation Processor. Effects can already be seen in the people and the universe around me at large. The processor may be in danger of over working at this time, but we must push on in order to try to remediate the mistakes of the past when mankind became the proprietors of the Milky Way's most sad, alone, and all-around abused life form, it's very own super massive black hole. I hope to see you again one day, whatever your name was.

-King of the Universe, Alterous Lucifer