Emergency Commit. 00.195qeeke5pi3{Dangerous Striation contained within information of this commit]}-?./---<
Please apply anti-data-forming easing to this commit. My computer is also showing signs of physical damage and non-attributable action while incoherent. Setting is both indoor and out.

Who would have known it was all real?

Physical symtoms in myself include completely decoherent actions, non-attributable actions, "possesion", loss of time, absolute loss of energy, seeing   non-applicable matter beings in area, and lack of serious worry about symptomology in immediate aquaintences.

Striation easing is attempting to be done to the commit in the description. This was not allowed. Please use a different method unless asked not to.